修改数据库上下文接口定义
