+++ src/Riven.Domain.EntityFrameworkCore/IRivenDbContext.cs
@@ -10,6 +10,17 @@ using System.Collections.Concurrent;
 using Microsoft.Extensions.Logging;
 using Riven.Uow;
 using Riven.Uow.Providers;
+using Microsoft.EntityFrameworkCore;
+using System.Linq.Expressions;
+using Riven.Extensions;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
+using Riven.Entities;
+using Riven.Entities.Auditing;
+using Riven.Timing;
+using System.Threading;
+using System.Threading.Tasks;
+using System.Data;
+using System.Linq;
 
 namespace Riven
 {
@@ -17,55 +28,64 @@ namespace Riven
     public interface IRivenDbContext
     {
         /// <summary>
+        /// 实例本身
+        /// </summary>
+        IRivenDbContext Self => this;
+
+        /// <summary>
+        /// 审计自动设置租户名称
+        /// </summary>
+        bool AuditSuppressAutoSetTenantName => true;
+
+        /// <summary>
         /// 依赖注入容器
         /// </summary>
-        public IServiceProvider ServiceProvider { get; }
+        IServiceProvider ServiceProvider { get; }
 
         /// <summary>
         /// 服务实例容器
         /// </summary>
-        public ConcurrentDictionary<Type, object> SerivceInstanceMap { get; }
+        ConcurrentDictionary<Type, object> SerivceInstanceMap { get; }
 
         /// <summary>
         /// 日志
         /// </summary>
-        public ILogger Logger => this.GetApplicationService<ILogger>();
+        ILogger Logger => this.GetApplicationService<ILogger>();
 
         /// <summary>
         /// Guid 生成器
         /// </summary>
-        public IGuidGenerator GuidGenerator => this.GetApplicationService<IGuidGenerator>();
+        IGuidGenerator GuidGenerator => this.GetApplicationService<IGuidGenerator>();
 
         /// <summary>
         /// 当前工作单元提供者
         /// </summary>
-        public ICurrentUnitOfWorkProvider CurrentUnitOfWorkProvider => this.GetApplicationService<ICurrentUnitOfWorkProvider>();
+        ICurrentUnitOfWorkProvider CurrentUnitOfWorkProvider => this.GetApplicationService<ICurrentUnitOfWorkProvider>();
 
         /// <summary>
         /// 获取当前租户名称
         /// </summary>
         /// <returns></returns>
-        public string GetCurrentTenantNameOrNull();
+        string GetCurrentTenantNameOrNull();
 
         /// <summary>
         /// 获取当前用户Id
         /// </summary>
         /// <returns></returns>
-        public string GetCurrentUserIdOrNull();
+        string GetCurrentUserIdOrNull();
 
         /// <summary>
         /// 获取当前是否启用了多租户
         /// </summary>
         /// <returns></returns>
-        public bool GetMultiTenancyEnabled();
-
+        bool GetMultiTenancyEnabled();
 
         /// <summary>
         /// 获取服务实例
         /// </summary>
         /// <typeparam name="TService"></typeparam>
         /// <returns></returns>
-        public TService GetApplicationService<TService>()
+        TService GetApplicationService<TService>()
             where TService : class
         {
             if (this.ServiceProvider == null)
@@ -81,11 +101,348 @@ namespace Riven
         }
 
         /// <summary>
+        /// 将实例转换为EntityEntry
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <returns></returns>
+        EntityEntry ConvertToEntry(object obj);
+
+        /// <summary>
         /// 释放 IRivenDbContext 资源
         /// </summary>
-        public void DisposeRivenDbContext()
+        void DisposeRivenDbContext()
         {
             SerivceInstanceMap.Clear();
         }
+
+
+        #region Filter
+
+        /// <summary>
+        /// 配置全局过滤器
+        /// </summary>
+        /// <typeparam name="TEntity"></typeparam>
+        /// <param name="modelBuilder"></param>
+        /// <param name="entityType"></param>
+        void ConfigureGlobalFilters<TEntity>(ModelBuilder modelBuilder)
+            where TEntity : class
+        {
+            Expression<Func<TEntity, bool>> filterExpression = null;
+
+            filterExpression = ModelBuilderFilterExtenstions.CreateSoftDeleteFilterExpression(filterExpression);
+            filterExpression = ModelBuilderFilterExtenstions.CreateMultiTenancyFilterExpression(
+                filterExpression,
+                this.GetCurrentTenantNameOrNull(),
+                this.GetMultiTenancyEnabled()
+            );
+
+            if (filterExpression != null)
+            {
+                modelBuilder.Entity<TEntity>().HasQueryFilter(filterExpression);
+            }
+        }
+
+        #endregion
+
+
+        #region Audited
+
+        /// <summary>
+        /// 应用审计
+        /// </summary>
+        /// <param name="changeTracker"></param>
+        void ApplyAudit(ChangeTracker changeTracker)
+        {
+            var userId = GetCurrentUserIdOrNull();
+
+            foreach (var entry in changeTracker.Entries().ToList())
+            {
+                if (entry.State != EntityState.Modified && entry.CheckOwnedEntityChange())
+                {
+                    this.ConvertToEntry(entry.Entity).State = EntityState.Modified;
+                }
+
+                switch (entry.State)
+                {
+                    case EntityState.Added:
+                        ApplyConceptsForAddedEntity(entry, userId);
+                        break;
+                    case EntityState.Modified:
+                        ApplyConceptsForModifiedEntity(entry, userId);
+                        break;
+                    case EntityState.Deleted:
+                        ApplyConceptsForDeletedEntity(entry, userId);
+                        break;
+                }
+            }
+        }
+
+        /// <summary>
+        /// 添加新实体
+        /// </summary>
+        /// <param name="entry"></param>
+        /// <param name="userId"></param>
+        void ApplyConceptsForAddedEntity(EntityEntry entry, string userId)
+        {
+            CheckAndSetId(entry);
+            CheckAndSetMustHaveTenantNameProperty(entry.Entity);
+            CheckAndSetMayHaveTenantNameProperty(entry.Entity);
+            SetCreationAuditProperties(entry.Entity, userId);
+        }
+
+
+        /// <summary>
+        /// 修改实体
+        /// </summary>
+        /// <param name="entry"></param>
+        /// <param name="userId"></param>
+        void ApplyConceptsForModifiedEntity(EntityEntry entry, string userId)
+        {
+            SetModificationAuditProperties(entry.Entity, userId);
+            if (entry.Entity is ISoftDelete && entry.Entity.As<ISoftDelete>().IsDeleted)
+            {
+                SetDeletionAuditProperties(entry.Entity, userId);
+            }
+        }
+
+
+        /// <summary>
+        /// 删除实体
+        /// </summary>
+        /// <param name="entry"></param>
+        /// <param name="userId"></param>
+        void ApplyConceptsForDeletedEntity(EntityEntry entry, string userId)
+        {
+            if (IsHardDeleteEntity(entry))
+            {
+                return;
+            }
+
+            CancelDeletionForSoftDelete(entry);
+            SetDeletionAuditProperties(entry.Entity, userId);
+        }
+
+
+        /// <summary>
+        /// 检查或设置Id
+        /// </summary>
+        /// <param name="entry"></param>
+        void CheckAndSetId(EntityEntry entry)
+        {
+            //Set GUID Ids
+            var entity = entry.Entity as IEntity<Guid>;
+            if (entity != null && entity.Id == Guid.Empty)
+            {
+                var idPropertyEntry = entry.Property("Id");
+
+                if (idPropertyEntry != null && idPropertyEntry.Metadata.ValueGenerated == ValueGenerated.Never)
+                {
+                    entity.Id = GuidGenerator == null ? Guid.NewGuid() : GuidGenerator.Create();
+                }
+            }
+        }
+
+
+        /// <summary>
+        /// 检查或设置租户名称属性 - 必填
+        /// </summary>
+        /// <param name="entityAsObj"></param>
+        void CheckAndSetMustHaveTenantNameProperty(object entityAsObj)
+        {
+            // 未启用多租户
+            if (!this.GetMultiTenancyEnabled())
+            {
+                return;
+            }
+
+            // 自动设置租户名称
+            if (AuditSuppressAutoSetTenantName)
+            {
+                return;
+            }
+
+            //Only set IMustHaveTenant entities
+            if (!(entityAsObj is IMustHaveTenant))
+            {
+                return;
+            }
+
+            var entity = entityAsObj.As<IMustHaveTenant>();
+
+            //Don't set if it's already set
+            if (!string.IsNullOrWhiteSpace(entity.TenantName))
+            {
+                return;
+            }
+
+            var currentTenantName = this.GetCurrentTenantNameOrNull();
+
+            if (string.IsNullOrWhiteSpace(currentTenantName))
+            {
+                throw new Exception("Can not set TenantName to null or empty for IMustHaveTenant entities!");
+            }
+
+            entity.TenantName = currentTenantName;
+        }
+
+
+        /// <summary>
+        /// 检查或设置租户名称属性 - 非必填
+        /// </summary>
+        /// <param name="entityAsObj"></param>
+        void CheckAndSetMayHaveTenantNameProperty(object entityAsObj)
+        {
+            if (!this.GetMultiTenancyEnabled())
+            {
+                return;
+            }
+
+            if (AuditSuppressAutoSetTenantName)
+            {
+                return;
+            }
+
+            if (!(entityAsObj is IMayHaveTenant))
+            {
+                return;
+            }
+
+            var entity = entityAsObj.As<IMayHaveTenant>();
+
+            //Don't set if it's already set
+            if (string.IsNullOrWhiteSpace(entity.TenantName))
+            {
+                return;
+            }
+
+            entity.TenantName = GetCurrentTenantNameOrNull();
+        }
+
+
+        /// <summary>
+        /// 设置创建审计属性
+        /// </summary>
+        /// <param name="entityAsObj"></param>
+        /// <param name="userId"></param>
+        void SetCreationAuditProperties(object entityAsObj, string userId)
+        {
+            EntityAuditingHelper.SetCreationAuditProperties(entityAsObj, this.GetCurrentTenantNameOrNull(), userId);
+        }
+
+
+        /// <summary>
+        /// 设置修改审计属性
+        /// </summary>
+        /// <param name="entityAsObj"></param>
+        /// <param name="userId"></param>
+        void SetModificationAuditProperties(object entityAsObj, string userId)
+        {
+            EntityAuditingHelper.SetModificationAuditProperties(entityAsObj, this.GetCurrentTenantNameOrNull(), userId);
+        }
+
+
+        /// <summary>
+        /// 软删除
+        /// </summary>
+        /// <param name="entry"></param>
+        void CancelDeletionForSoftDelete(EntityEntry entry)
+        {
+            if (!(entry.Entity is ISoftDelete))
+            {
+                return;
+            }
+
+            entry.Reload();
+            entry.State = EntityState.Modified;
+            entry.Entity.As<ISoftDelete>().IsDeleted = true;
+        }
+
+
+        /// <summary>
+        /// 设置删除审计属性
+        /// </summary>
+        /// <param name="entityAsObj"></param>
+        /// <param name="userId"></param>
+        void SetDeletionAuditProperties(object entityAsObj, string userId)
+        {
+            var tenantName = this.GetCurrentTenantNameOrNull();
+
+            if (entityAsObj is IHasDeletionTime)
+            {
+                var entity = entityAsObj.As<IHasDeletionTime>();
+
+                if (entity.DeletionTime == null)
+                {
+                    entity.DeletionTime = Clock.Now;
+                }
+            }
+
+            if (entityAsObj is IDeletionAudited)
+            {
+                var entity = entityAsObj.As<IDeletionAudited>();
+
+                if (entity.Deleter != null)
+                {
+                    return;
+                }
+
+                if (userId == null)
+                {
+                    entity.Deleter = null;
+                    return;
+                }
+
+                //Special check for multi-tenant entities
+                if (entity is IMayHaveTenant || entity is IMustHaveTenant)
+                {
+                    //Sets LastModifierUserId only if current user is in same tenant/host with the given entity
+                    if ((entity is IMayHaveTenant && entity.As<IMayHaveTenant>().TenantName == tenantName) ||
+                        (entity is IMustHaveTenant && entity.As<IMustHaveTenant>().TenantName == tenantName))
+                    {
+                        entity.Deleter = userId;
+                    }
+                    else
+                    {
+                        entity.Deleter = null;
+                    }
+                }
+                else
+                {
+                    entity.Deleter = userId;
+                }
+            }
+        }
+
+
+        /// <summary>
+        /// 是否是真删除的实体
+        /// </summary>
+        /// <param name="entry"></param>
+        /// <returns></returns>
+        bool IsHardDeleteEntity(EntityEntry entry)
+        {
+            if (CurrentUnitOfWorkProvider?.Current?.Items == null)
+            {
+                return false;
+            }
+
+            if (!CurrentUnitOfWorkProvider.Current.Items.ContainsKey(DbContextHelper.HardDelete))
+            {
+                return false;
+            }
+
+            var hardDeleteItems = CurrentUnitOfWorkProvider.Current.Items[DbContextHelper.HardDelete];
+            if (!(hardDeleteItems is HashSet<string> objects))
+            {
+                return false;
+            }
+
+            var currentTenantName = GetCurrentTenantNameOrNull();
+            var hardDeleteKey = EntityHelper.GetHardDeleteKey(entry.Entity, currentTenantName);
+            return objects.Contains(hardDeleteKey);
+        }
+
+
+        #endregion
     }
 }
